fix(utils): harden owned user upload URL detection
Match Vercel Blob host/https rules used for org and coworker images so
custom avatar protection cannot be spoofed by foreign hosts.
