+++ packages/utils/src/__tests__/user-upload-path.test.ts
@@ -3,7 +3,6 @@ import { describe, expect, it } from "vitest";
 import {
   buildUserUploadPathname,
   buildUserUploadPrefix,
-  isOwnedUserUploadUrl,
   sanitizeUserUploadFilename,
 } from "../user-upload-path.js";
 
@@ -26,36 +25,5 @@ describe("user upload path helpers", () => {
     expect(buildUserUploadPathname("user_123", "hello world.txt")).toBe(
       "users/user_123/hello_world.txt",
     );
-  });
-
-  it("detects owned user upload blob URLs", () => {
-    expect(
-      isOwnedUserUploadUrl(
-        "https://blob.example/users/user_123/file.png",
-        "user_123",
-      ),
-    ).toBe(true);
-  });
-
-  it("rejects uploads owned by a different user", () => {
-    expect(
-      isOwnedUserUploadUrl(
-        "https://blob.example/users/user_999/file.png",
-        "user_123",
-      ),
-    ).toBe(false);
-  });
-
-  it("rejects non-upload and invalid URLs", () => {
-    expect(
-      isOwnedUserUploadUrl(
-        "https://lh3.googleusercontent.com/a/xyz",
-        "user_123",
-      ),
-    ).toBe(false);
-    expect(isOwnedUserUploadUrl("not-a-url", "user_123")).toBe(false);
-    expect(
-      isOwnedUserUploadUrl("https://blob.example/users/user_123/file.png", ""),
-    ).toBe(false);
   });
 });
