+++ packages/utils/src/index.ts
@@ -244,10 +244,10 @@ export {
   USER_UPLOAD_ALLOWED_CONTENT_TYPE_SET,
   USER_UPLOAD_ALLOWED_CONTENT_TYPES,
 } from "./user-upload-content-type.js";
+export { isOwnedUserUploadUrl } from "./user-upload-owned.js";
 export {
   buildUserUploadPathname,
   buildUserUploadPrefix,
-  isOwnedUserUploadUrl,
   sanitizeUserUploadFilename,
 } from "./user-upload-path.js";
 export {
